# web/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Sieve API",
    "summary": "Evidence-backed dependency vulnerability triage.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/repositories/{repository_id}": {
      "get": {
        "tags": [
          "repositories"
        ],
        "summary": "Repository",
        "operationId": "repository_api_v1_repositories__repository_id__get",
        "parameters": [
          {
            "name": "repository_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Repository Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RepositoryOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/repositories/{repository_id}/scans": {
      "get": {
        "tags": [
          "repositories"
        ],
        "summary": "Scans",
        "operationId": "scans_api_v1_repositories__repository_id__scans_get",
        "parameters": [
          {
            "name": "repository_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Repository Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ScanOut"
                  },
                  "title": "Response Scans Api V1 Repositories  Repository Id  Scans Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "repositories"
        ],
        "summary": "Rescan",
        "operationId": "rescan_api_v1_repositories__repository_id__scans_post",
        "parameters": [
          {
            "name": "repository_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Repository Id"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScanOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/repositories/{repository_id}/pull-requests": {
      "get": {
        "tags": [
          "repositories"
        ],
        "summary": "Pull Requests",
        "operationId": "pull_requests_api_v1_repositories__repository_id__pull_requests_get",
        "parameters": [
          {
            "name": "repository_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Repository Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PullRequestCheckOut"
                  },
                  "title": "Response Pull Requests Api V1 Repositories  Repository Id  Pull Requests Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/repositories/{repository_id}/vex/preview": {
      "get": {
        "tags": [
          "repositories"
        ],
        "summary": "Vex Preview",
        "description": "The document as it would be generated now. Nothing is stored. Available to every viewer,\nincluding demo visitors.",
        "operationId": "vex_preview_api_v1_repositories__repository_id__vex_preview_get",
        "parameters": [
          {
            "name": "repository_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Repository Id"
            }
          },
          {
            "name": "format",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/VexFormat",
              "default": "openvex"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Vex Preview Api V1 Repositories  Repository Id  Vex Preview Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/repositories/{repository_id}/vex": {
      "post": {
        "tags": [
          "repositories"
        ],
        "summary": "Create Vex",
        "operationId": "create_vex_api_v1_repositories__repository_id__vex_post",
        "parameters": [
          {
            "name": "repository_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Repository Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VexIn"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VexDocumentOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PullRequestCheckOut": {
        "properties": {
          "pr_number": {
            "type": "integer",
            "title": "Pr Number"
          },
          "head_sha": {
            "type": "string",
            "title": "Head Sha"
          },
          "conclusion": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Conclusion"
          },
          "summary": {
            "additionalProperties": true,
            "type": "object",
            "title": "Summary"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "pr_number",
          "head_sha",
          "conclusion",
          "summary",
          "created_at"
        ],
        "title": "PullRequestCheckOut"
      },
...
      "RepositoryOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "full_name": {
            "type": "string",
            "title": "Full Name"
          },
          "source": {
            "type": "string",
            "title": "Source"
          },
          "default_branch": {
            "type": "string",
            "title": "Default Branch"
          },
          "is_private": {
            "type": "boolean",
            "title": "Is Private"
          },
          "latest_scan": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ScanOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "counts": {
            "$ref": "#/components/schemas/VerdictCounts"
          },
          "health": {
            "type": "string",
            "enum": [
              "healthy",
              "attention",
              "failing",
              "pending"
            ],
            "title": "Health"
          }
        },
        "type": "object",
        "required": [
          "id",
          "full_name",
          "source",
          "default_branch",
          "is_private",
          "latest_scan",
          "counts",
          "health"
        ],
        "title": "RepositoryOut"
      },
...
      "ScanOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "repository_id": {
            "type": "string",
            "format": "uuid",
            "title": "Repository Id"
          },
          "commit_sha": {
            "type": "string",
            "title": "Commit Sha"
          },
          "ref": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ref"
          },
          "trigger": {
            "$ref": "#/components/schemas/ScanTrigger"
          },
          "status": {
            "$ref": "#/components/schemas/ScanStatus"
          },
          "stage": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stage"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "finished_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "stats": {
            "additionalProperties": true,
            "type": "object",
            "title": "Stats"
          },
          "error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Code"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          }
        },
        "type": "object",
        "required": [
          "id",
          "repository_id",
          "commit_sha",
          "ref",
          "trigger",
          "status",
          "stage",
          "created_at",
          "started_at",
          "finished_at",
          "stats",
          "error_code",
          "error_message"
        ],
        "title": "ScanOut"
      },
      "ScanStatus": {
        "type": "string",
        "enum": [
          "queued",
          "running",
          "succeeded",
          "failed",
          "cancelled"
        ],
        "title": "ScanStatus"
      },
      "ScanTrigger": {
        "type": "string",
        "enum": [
          "manual",
          "push",
          "pull_request",
          "advisory",
          "schedule",
          "demo"
        ],
        "title": "ScanTrigger"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "VerdictCounts": {
        "properties": {
          "total": {
            "type": "integer",
            "title": "Total",
            "default": 0
          },
          "reachable": {
            "type": "integer",
            "title": "Reachable",
            "default": 0
          },
          "needs_review": {
            "type": "integer",
            "title": "Needs Review",
            "default": 0
          },
          "not_reached": {
            "type": "integer",
            "title": "Not Reached",
            "default": 0
          },
          "pending": {
            "type": "integer",
            "title": "Pending",
            "default": 0
          },
          "kev": {
            "type": "integer",
            "title": "Kev",
            "default": 0
          }
        },
        "type": "object",
        "title": "VerdictCounts"
      },
      "VexDocumentOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "format": {
            "$ref": "#/components/schemas/VexFormat"
          },
          "repository_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Repository Id"
          },
          "repository_full_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Repository Full Name"
          },
          "statement_count": {
            "type": "integer",
            "title": "Statement Count"
          },
          "sha256": {
            "type": "string",
            "title": "Sha256"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "format",
          "repository_id",
          "repository_full_name",
          "statement_count",
          "sha256",
          "created_at"
        ],
        "title": "VexDocumentOut"
      },
      "VexFormat": {
        "type": "string",
        "enum": [
          "openvex",
          "cyclonedx"
        ],
        "title": "VexFormat"
      },
      "VexIn": {
        "properties": {
          "format": {
            "$ref": "#/components/schemas/VexFormat"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "format"
        ],
        "title": "VexIn"
      },
...
    }
  }
}